+++ src/main/java/classfit/example/classfit/mail/service/EmailAsyncService.java
@@ -1,0 +1,39 @@
+package classfit.example.classfit.mail.service;
+
+import jakarta.mail.MessagingException;
+import jakarta.mail.internet.MimeMessage;
+import lombok.RequiredArgsConstructor;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Service;
+import org.thymeleaf.context.Context;
+import org.thymeleaf.spring6.SpringTemplateEngine;
+
+import java.util.Map;
+
+@Service
+@RequiredArgsConstructor
+public class EmailAsyncService {
+
+    private final JavaMailSender javaMailSender;
+    private final SpringTemplateEngine templateEngine;
+
+    @Async("emailAsyncExecutor")
+    public void sendEmailForm(String email, String subject, Map<String, Object> templateVariables) {
+        try {
+            MimeMessage message = javaMailSender.createMimeMessage();
+            message.addRecipients(MimeMessage.RecipientType.TO, email);
+            message.setSubject(subject);
+
+            Context context = new Context();
+            context.setVariables(templateVariables);
+
+            String processedTemplate = templateEngine.process("mail", context);
+            message.setText(processedTemplate, "utf-8", "html");
+
+            javaMailSender.send(message);
+        } catch (MessagingException e) {
+            throw new RuntimeException("이메일 전송 실패");
+        }
+    }
+}
